Standard issue jlooch
This project comes from not wanting to install the jsyn plugin for my browser.
JLooch expects to find it or something, so I'm going to change its defaults
to work in exactly this case. Further, jsyn is a noncommercial license, so I
will refrain from putting this on github. Jlooch's creator is license agnostic.

So, at this point, everything is standard, exactly as downloaded. I did, however, unpack the jsyn jar and put in blah.html . I'd much rather have a swing thingy than an applet, but whatever. That's for a different branch.